fixed sle side

+++ server/routes/calendar.js
@@ -37,28 +37,32 @@ router.post("/changecoverage", (req, res) => {
   return res.json({ Successful: true });
 });
 router.post("/save", (req, res) => {
-  items = req.body.items;
-  var userId = req.body.userId;
-  pool.query(
-    "DELETE FROM AVAILABILITY WHERE sle_id=$1",
-    [userId],
-    (error, result) => {
-      if (error) {
-        throw error;
-      }
-    }
-  );
-  for (var i = 0; i < items.length; i += 1) {
+  if (!req.user) {
+    return res.json({ schedule: null });
+  } else {
+    items = req.body.items;
+    var userId = req.user;
     pool.query(
-      `INSERT INTO AVAILABILITY (sle_id, start_time, day_of_week) VALUES (${userId}, ${items[i][0]}, ${items[i][1]})`,
+      "DELETE FROM AVAILABILITY WHERE sle_id=$1",
+      [userId],
       (error, result) => {
         if (error) {
           throw error;
         }
       }
     );
+    for (var i = 0; i < items.length; i += 1) {
+      pool.query(
+        `INSERT INTO AVAILABILITY (sle_id, start_time, day_of_week) VALUES (${userId}, ${items[i][0]}, ${items[i][1]})`,
+        (error, result) => {
+          if (error) {
+            throw error;
+          }
+        }
+      );
+    }
+    return res.json({ schedule: items });
   }
-  return res.json({ schedule: items });
 });
 
 router.post("/staticcalendar", (req, res) => {
@@ -130,38 +134,42 @@ router.get("/shifts", function (req, res) {
   });
 });
 
-router.get("/availability/:userId", (req, res) => {
+router.get("/availability", (req, res) => {
   var selected = [];
   var curr_day = new Date();
   var curr_week_sunday = curr_day.getDate() - curr_day.getDay();
-  pool.query(
-    `SELECT start_time AS t, day_of_week AS d FROM AVAILABILITY 
+  if (!req.user) {
+    return res.json({ schedule: null });
+  } else {
+    pool.query(
+      `SELECT start_time AS t, day_of_week AS d FROM AVAILABILITY 
      WHERE sle_id = $1`,
-    [req.params.userId],
-    (error, result) => {
-      if (error) {
-        throw error;
-      } else {
-        for (var r = 0; r < result.rows.length; r++) {
-          var row = result.rows[r];
-          var t = result.rows[r].t;
-          var d = result.rows[r].d;
-          selected.push(
-            new Date(
-              curr_day.getFullYear(),
-              curr_day.getMonth(),
-              d + curr_week_sunday,
-              t,
-              0,
-              0,
-              0
-            )
-          );
+      [req.user],
+      (error, result) => {
+        if (error) {
+          throw error;
+        } else {
+          for (var r = 0; r < result.rows.length; r++) {
+            var row = result.rows[r];
+            var t = result.rows[r].t;
+            var d = result.rows[r].d;
+            selected.push(
+              new Date(
+                curr_day.getFullYear(),
+                curr_day.getMonth(),
+                d + curr_week_sunday,
+                t,
+                0,
+                0,
+                0
+              )
+            );
+          }
         }
+        return res.json({ schedule: selected });
       }
-      return res.json({ schedule: selected });
-    }
-  );
+    );
+  }
 });
 router.get("/totalhours/:userId", (req, res) => {
   //Accouting for daylight savings
